compile again

+++ src/lot.rs
@@ -59,7 +59,7 @@ impl Lot {
         self.state
     }
 
-    pub fn get_id(&self) -> u32 {
+    pub fn get_id(&self) -> i64 {
         self.id
     }
